feature: add controllers for for agents

+++ src/main/java/io/eskay/ticket_system/service/TicketServiceImpl.java
@@ -6,13 +6,14 @@ import io.eskay.ticket_system.dto.request.UpdateTicketRequest;
 import io.eskay.ticket_system.dto.response.CommentDto;
 import io.eskay.ticket_system.dto.response.TicketDto;
 import io.eskay.ticket_system.entity.Ticket;
+import io.eskay.ticket_system.entity.TicketStatus;
+import io.eskay.ticket_system.exception.ForbiddenOperationException;
 import io.eskay.ticket_system.exception.ResourceNotFoundException;
 import io.eskay.ticket_system.mapper.CommentDtoMapper;
 import io.eskay.ticket_system.mapper.CommentRequestMapper;
 import io.eskay.ticket_system.mapper.TicketDtoMapper;
 import io.eskay.ticket_system.mapper.TicketRequestMapper;
 import io.eskay.ticket_system.repository.CategoryRepository;
-import io.eskay.ticket_system.repository.TicketCommentRepository;
 import io.eskay.ticket_system.repository.TicketRepository;
 import io.eskay.ticket_system.util.AuthenticatedUser;
 import org.springframework.stereotype.Service;
@@ -27,7 +28,6 @@ public class TicketServiceImpl implements TicketService {
     private final TicketDtoMapper ticketDtoMapper;
     private final TicketRequestMapper requestMapper;
     private final CategoryRepository categoryRepository;
-    private final TicketCommentRepository commentRepository;
     private final CommentDtoMapper commentDtoMapper;
     private final CommentRequestMapper commentRequestMapper;
 
@@ -36,14 +36,13 @@ public class TicketServiceImpl implements TicketService {
             TicketDtoMapper ticketDtoMapper,
             TicketRequestMapper requestMapper,
             CategoryRepository categoryRepository,
-            TicketCommentRepository commentRepository,
             CommentDtoMapper commentDtoMapper,
-            CommentRequestMapper commentRequestMapper) {
+            CommentRequestMapper commentRequestMapper
+    ) {
         this.ticketRepository = ticketRepository;
         this.ticketDtoMapper = ticketDtoMapper;
         this.requestMapper = requestMapper;
         this.categoryRepository = categoryRepository;
-        this.commentRepository = commentRepository;
         this.commentDtoMapper = commentDtoMapper;
         this.commentRequestMapper = commentRequestMapper;
     }
@@ -59,24 +58,21 @@ public class TicketServiceImpl implements TicketService {
 
     @Override
     public TicketDto getTicket(Long id) {
-        return ticketRepository
-                .findByIdAndRaisedBy(id, AuthenticatedUser.get())
-                .map(ticketDtoMapper)
-                .orElseThrow(() -> new ResourceNotFoundException(
-                        "User does not have a ticket with id '%d', check and try again".formatted(id)));
+        return ticketDtoMapper.apply(getTicketByIdAndRaisedBy(id));
     }
 
     @Override
     @Transactional
     public TicketDto createTicket(CreateTicketRequest request) {
         var category = categoryRepository
-                .findById(request.category())
-                .orElseThrow(() -> new ResourceNotFoundException(
+                .findByIdWithTeam(request.category())
+                .orElseThrow(() -> new IllegalArgumentException(
                         "Error specified category not found, check and try again"
                 ));
 
         var ticket = requestMapper.apply(request);
         ticket.setCategory(category);
+        ticket.setTeam(category.getDefaultTeam());
 
         return ticketDtoMapper.apply(ticketRepository.save(ticket));
     }
@@ -84,53 +80,68 @@ public class TicketServiceImpl implements TicketService {
     @Override
     @Transactional
     public TicketDto updateTicket(Long id, UpdateTicketRequest request) {
-        var foundTicket = ticketRepository
-                .findByIdAndRaisedBy(id, AuthenticatedUser.get())
-                .orElseThrow(() -> new ResourceNotFoundException(
-                        "User does not have a ticket with id '%d', check and try again".formatted(id)));
+        var foundTicket = getTicketByIdAndRaisedBy(id);
 
-        updateTicket(request, foundTicket);
+        updateTicketDetails(request, foundTicket);
         return ticketDtoMapper.apply(ticketRepository.save(foundTicket));
     }
 
     @Override
     @Transactional
     public void deleteTicket(Long id) {
-        ticketRepository
-                .findByIdAndRaisedBy(id, AuthenticatedUser.get())
-                .orElseThrow(() -> new ResourceNotFoundException(
-                        "User does not have a ticket with id '%d', check and try again".formatted(id)));
-        ticketRepository.deleteById(id);
+        var foundTicket = getTicketByIdAndRaisedBy(id);
+        if (
+                foundTicket.getStatus().equals(TicketStatus.PENDING) ||
+                foundTicket.getStatus().equals(TicketStatus.CLOSED)
+        ) {
+            ticketRepository.deleteById(id);
+            return;
+        }
+
+        throw new ForbiddenOperationException(
+                "Forbidden, ticket cannot be deleted whiles it is %s"
+                        .formatted(foundTicket.getStatus().name()));
     }
 
     @Override
     public List<CommentDto> getTicketComments(Long id) {
-        var foundTicket = ticketRepository
-                .findByIdAndRaisedBy(id, AuthenticatedUser.get())
-                .orElseThrow(() -> new ResourceNotFoundException(
-                        "User does not have a ticket with id '%d', check and try again".formatted(id)));
+        var foundTicket = getTicketByIdAndRaisedByWithComments(id);
 
-        return commentRepository
-                .findAllByAuthorAndTicket
-                        (AuthenticatedUser.get(), foundTicket)
+        return foundTicket
+                .getTicketComments()
                 .stream()
                 .map(commentDtoMapper)
                 .toList();
     }
 
     @Override
-    public CommentDto addComment(Long id, CreateCommentRequest request) {
-        var foundTicket = ticketRepository
+    @Transactional
+    public CommentDto addCommentToTicket(Long id, CreateCommentRequest request) {
+        var foundTicket = getTicketByIdAndRaisedByWithComments(id);
+
+        var comment = commentRequestMapper.apply(request);
+        foundTicket.addComment(comment);
+        ticketRepository.save(foundTicket);
+
+        return commentDtoMapper.apply(foundTicket.getTicketComments().getLast());
+    }
+
+
+    private Ticket getTicketByIdAndRaisedBy(Long id) {
+        return ticketRepository
                 .findByIdAndRaisedBy(id, AuthenticatedUser.get())
                 .orElseThrow(() -> new ResourceNotFoundException(
                         "User does not have a ticket with id '%d', check and try again".formatted(id)));
-
-        var comment = commentRequestMapper.apply(request);
-        comment.setTicket(foundTicket);
-        return commentDtoMapper.apply(commentRepository.save(comment));
     }
 
-    private void updateTicket(
+    private Ticket getTicketByIdAndRaisedByWithComments(Long id) {
+        return ticketRepository
+                .findByIdAndRaisedByWithComments(id, AuthenticatedUser.get())
+                .orElseThrow(() -> new ResourceNotFoundException(
+                        "User does not have a ticket with id '%d', check and try again".formatted(id)));
+    }
+
+    private void updateTicketDetails(
             UpdateTicketRequest request,
             Ticket foundTicket
     ) {
